backend/requirements.txt
flask
flask-cors
mtgsdk
psycopg2-binary
sqlalchemy
openai
python-dotenv
requests 